requirements.txt
# Core dependencies
PyYAML>=6.0
requests>=2.28.0
click>=8.1.0

# GitHub integration
PyGithub>=1.58.0
GitPython>=3.1.0

# Utilities
python-dateutil>=2.8.0
pathlib2>=2.3.0
typing-extensions>=4.3.0

# Optional dependencies for enhanced functionality
psutil>=5.9.0  # For system monitoring
colorama>=0.4.0  # For colored terminal output

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0